feat : application.yml 수정, 불필요 코드문 제거, 불필요 임포트 제거, setter제거, erd, 시퀀스 업로드, final 추가, id long 타입으로변경, @Column 어노테이션 추가

+++ src/main/java/kr/hhplus/be/server/domain/product/repository/ProductRepository.java
@@ -3,8 +3,6 @@ package kr.hhplus.be.server.domain.product.repository;
 import kr.hhplus.be.server.domain.product.entity.Product;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 
 import java.util.List;
 import java.util.Optional;
@@ -16,7 +14,7 @@ public interface ProductRepository {
     Page<Product> getProducts(Pageable pageable);
 
     //상품 1행 정보 조회
-    Optional<Product> findById(int id);
+    Optional<Product> findById(long id);
 
     //Top5
     Optional<List<Product>> getTop5List();
