+++ src/main/java/kr/hhplus/be/server/infra/point/jparepository/PointJpaRepository.java
@@ -3,7 +3,7 @@ package kr.hhplus.be.server.infra.point.jparepository;
 import kr.hhplus.be.server.domain.point.entity.Point;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PointJpaRepository extends JpaRepository<Point, Integer> {
+public interface PointJpaRepository extends JpaRepository<Point, Long> {
 
     //포인트 조회
     Point findByUserId(int userId);
